+++ Professor-X/src/main/java/ai/profX/service/GameLogService.java
@@ -3,5 +3,5 @@ package ai.profX.service;
 import java.util.HashMap;
 
 public interface GameLogService {
-	public void addGameLog(long charId, HashMap<Integer, Integer> log,Boolean result);
+	public void addGameLog(long charId, HashMap<Long, Integer> log,Boolean result);
 }
